chore: cleanup functions and docs
